requirements.txt
flask
gunicorn
uvicorn
google-genai
google-generativeai
